backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
supabase==2.3.4
python-dotenv==1.0.0
Pillow==10.0.1
Werkzeug==2.3.7
requests==2.31.0
APScheduler==3.10.4
pytz==2023.3
PyJWT==2.8.0
